Handle Individual, DataValue and VariableValue comparison with NaNs
